book _recomendations/requirements.txt
pandas
tqdm
python-dotenv
pinecone-client
google-generativeai
